Rename spring module to spring-boot-3
