+++ inertia4j.spring-boot-3/src/main/java/io/github/inertia4j/springboot3/InertiaConfigurationProperties.java
@@ -1,4 +1,4 @@
-package io.github.inertia4j.spring;
+package io.github.inertia4j.springboot3;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.bind.ConstructorBinding;
