+++ inertia4j.spring-boot-3/src/main/java/io/github/inertia4j/springboot3/InertiaSpringRenderer.java
@@ -1,4 +1,4 @@
-package io.github.inertia4j.spring;
+package io.github.inertia4j.springboot3;
 
 import io.github.inertia4j.core.HttpRequest;
 import io.github.inertia4j.core.HttpResponse;
