+++ inertia4j.spring-boot-3/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,2 @@
+io.github.inertia4j.springboot3.InertiaSpringAutoconfiguration
+io.github.inertia4j.springboot3.InertiaConfigurationProperties
