+++ inertia4j.spring-boot-3/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+spring.application.name=inertia4j-spring-boot-3
