+++ inertia4j.spring-boot-3/src/test/java/io/github/inertia4j/springboot3/InertiaMockMvcTest.java
@@ -1,4 +1,4 @@
-package io.github.inertia4j.spring;
+package io.github.inertia4j.springboot3;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
