+++ inertia4j.spring-boot-3/src/test/resources/templates/app.html
