+++ settings.gradle.kts
@@ -7,4 +7,4 @@ rootProject.name = "inertia4j"
 include("inertia4j.spi")
 include("inertia4j.core")
 include("inertia4j.ktor")
-include("inertia4j.spring")
+include("inertia4j.spring-boot-3")
